+++ contract-macro/tests/compile-pass/src/methods/mod.rs
@@ -1,3 +1,4 @@
+pub mod deploy_constructor_exports_as_init;
 pub mod init_absent;
 pub mod init_no_params;
 pub mod new_returns_contract_name;